feat: integrate nutrition API

+++ app/src/main/java/com/zihanwang/myhealth/presentation/screen/nutrition/db/FoodEntity.kt
@@ -12,7 +12,7 @@ data class FoodEntity(
     val carb: Float,
     val protein: Float,
     val fat: Float,
-    val sodium: Int,
+    val sodium: Float,
     val sugar: Float,
     val satFat: Float
 )
